migrate hibernatesearch to indy
