+++ apm-agent-plugins/apm-hibernate-search-plugin/apm-hibernate-search-plugin-5_x/src/test/java/co/elastic/apm/agent/hibernatesearch/EntityManagerFactoryHelper.java
@@ -11,9 +11,9 @@
  * the Apache License, Version 2.0 (the "License"); you may
  * not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -22,7 +22,7 @@
  * under the License.
  * #L%
  */
-package co.elastic.apm.agent.hibernate.search.v6_x;
+package co.elastic.apm.agent.hibernatesearch;
 
 import java.nio.file.Path;
 import java.util.HashMap;
@@ -35,10 +35,7 @@ public class EntityManagerFactoryHelper {
 
     public static EntityManagerFactory buildEntityManagerFactory(final Path tempDirectory) {
         Map<String, Object> configOverrides = new HashMap<>();
-        configOverrides.put("hibernate.search.backends.testBackend.type", "lucene");
-        configOverrides.put("hibernate.search.backends.testBackend.directory_provider", "local_directory");
-        configOverrides.put("hibernate.search.backends.testBackend.root_directory", tempDirectory.toAbsolutePath().toString());
-        configOverrides.put("hibernate.search.default_backend", "testBackend");
+        configOverrides.put("hibernate.search.default.indexBase", tempDirectory.toAbsolutePath().toString());
         return Persistence.createEntityManagerFactory("templatePU", configOverrides);
     }
 }
