+++ apm-agent-plugins/apm-hibernate-search-plugin/apm-hibernate-search-plugin-6_x/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,1 +1,1 @@
-co.elastic.apm.agent.hibernate.search.v6_x.HibernateSearch6Instrumentation
+co.elastic.apm.agent.hibernatesearch.HibernateSearch6Instrumentation
